+++ packages/dash-table/tests/cypress/tests/unit/single_column_syntactic_tree_test.ts
@@ -1,10 +1,17 @@
 import { SingleColumnSyntaxTree } from 'dash-table/syntax-tree';
 import { ColumnType } from 'dash-table/components/Table/props';
 import { SingleColumnConfig } from 'dash-table/syntax-tree/SingleColumnSyntaxTree';
+import { RelationalOperator } from 'dash-table/syntax-tree/lexeme/relational';
+import { LexemeType } from 'core/syntax-tree/lexicon';
 
 const COLUMN_ANY: SingleColumnConfig = {
     id: 'a',
     type: ColumnType.Any
+};
+
+const COLUMN_DATE: SingleColumnConfig = {
+    id: 'a',
+    type: ColumnType.Datetime
 };
 
 const COLUMN_NUMERIC: SingleColumnConfig = {
@@ -72,7 +79,9 @@ describe('Single Column Syntax Tree', () => {
         const tree = new SingleColumnSyntaxTree('1', COLUMN_UNDEFINED);
 
         expect(tree.isValid).to.equal(true);
-        expect(tree.evaluate({ a: 1 })).to.equal(true);
+        expect(tree.evaluate({ a: '1' })).to.equal(true);
+        expect(tree.evaluate({ a: '2' })).to.equal(false);
+        expect(tree.evaluate({ a: 1 })).to.equal(false);
         expect(tree.evaluate({ a: 2 })).to.equal(false);
 
         expect(tree.toQueryString()).to.equal('{a} contains 1');
@@ -110,9 +119,119 @@ describe('Single Column Syntax Tree', () => {
         const tree = new SingleColumnSyntaxTree('"1"', COLUMN_TEXT);
 
         expect(tree.isValid).to.equal(true);
+        expect(tree.evaluate({ a: 1 })).to.equal(true);
+        expect(tree.evaluate({ a: 2 })).to.equal(false);
         expect(tree.evaluate({ a: '1' })).to.equal(true);
         expect(tree.evaluate({ a: '2' })).to.equal(false);
 
         expect(tree.toQueryString()).to.equal('{a} contains "1"');
+    });
+
+    ['1975', '"1975"'].forEach(value => {
+        it(`can be expression '${value}' with datetime column type`, () => {
+            const tree = new SingleColumnSyntaxTree(value, COLUMN_DATE);
+
+            expect(tree.evaluate({ a: 1975 })).to.equal(true);
+            expect(tree.evaluate({ a: '1975' })).to.equal(true);
+            expect(tree.evaluate({ a: '1975-01' })).to.equal(true);
+            expect(tree.evaluate({ a: '1975-01-01' })).to.equal(true);
+            expect(tree.evaluate({ a: '1975-01-01 01:01:01' })).to.equal(true);
+
+            expect(tree.evaluate({ a: 1976 })).to.equal(false);
+            expect(tree.evaluate({ a: '1976' })).to.equal(false);
+            expect(tree.evaluate({ a: '1976-01' })).to.equal(false);
+            expect(tree.evaluate({ a: '1976-01-01' })).to.equal(false);
+            expect(tree.evaluate({ a: '1976-01-01 01:01:01' })).to.equal(false);
+        });
+    });
+
+    [
+        { type: COLUMN_UNDEFINED, name: 'undefined' },
+        { type: COLUMN_ANY, name: 'any' },
+        { type: COLUMN_TEXT, name: 'text' }
+    ].forEach(({ type, name }) => {
+        it(`returns the correct relational operator lexeme for '${name}' column type`, () => {
+            const tree = new SingleColumnSyntaxTree('1', type);
+            const structure = tree.toStructure();
+
+            expect(tree.toQueryString()).to.equal('{a} contains 1');
+            expect(structure).to.not.equal(null);
+
+            if (structure) {
+                expect(structure.value).to.equal(RelationalOperator.Contains);
+                expect(structure.subType).to.equal(RelationalOperator.Contains);
+                expect(structure.type).to.equal(LexemeType.RelationalOperator);
+
+                expect(structure.left).to.not.equal(null);
+                if (structure.left) {
+                    expect(structure.left.type).to.equal(LexemeType.Expression);
+                    expect(structure.left.subType).to.equal('field');
+                    expect(structure.left.value).to.equal('a');
+                }
+
+                expect(structure.right).to.not.equal(null);
+                if (structure.right) {
+                    expect(structure.right.type).to.equal(LexemeType.Expression);
+                    expect(structure.right.subType).to.equal('value');
+                    expect(structure.right.value).to.equal(1);
+                }
+            }
+        });
+    });
+
+    it(`returns the correct relational operator lexeme for 'date' column type`, () => {
+        const tree = new SingleColumnSyntaxTree('1975', COLUMN_DATE);
+        const structure = tree.toStructure();
+
+        expect(tree.toQueryString()).to.equal('{a} datestartswith 1975');
+        expect(structure).to.not.equal(null);
+
+        if (structure) {
+            expect(structure.value).to.equal(RelationalOperator.DateStartsWith);
+            expect(structure.subType).to.equal(RelationalOperator.DateStartsWith);
+            expect(structure.type).to.equal(LexemeType.RelationalOperator);
+
+            expect(structure.left).to.not.equal(null);
+            if (structure.left) {
+                expect(structure.left.type).to.equal(LexemeType.Expression);
+                expect(structure.left.subType).to.equal('field');
+                expect(structure.left.value).to.equal('a');
+            }
+
+            expect(structure.right).to.not.equal(null);
+            if (structure.right) {
+                expect(structure.right.type).to.equal(LexemeType.Expression);
+                expect(structure.right.subType).to.equal('value');
+                expect(structure.right.value).to.equal(1975);
+            }
+        }
+    });
+
+    it(`returns the correct relational operator lexeme for 'numeric' column type`, () => {
+        const tree = new SingleColumnSyntaxTree('1', COLUMN_NUMERIC);
+        const structure = tree.toStructure();
+
+        expect(tree.toQueryString()).to.equal('{a} = 1');
+        expect(structure).to.not.equal(null);
+
+        if (structure) {
+            expect(structure.value).to.equal(RelationalOperator.Equal);
+            expect(structure.subType).to.equal(RelationalOperator.Equal);
+            expect(structure.type).to.equal(LexemeType.RelationalOperator);
+
+            expect(structure.left).to.not.equal(null);
+            if (structure.left) {
+                expect(structure.left.type).to.equal(LexemeType.Expression);
+                expect(structure.left.subType).to.equal('field');
+                expect(structure.left.value).to.equal('a');
+            }
+
+            expect(structure.right).to.not.equal(null);
+            if (structure.right) {
+                expect(structure.right.type).to.equal(LexemeType.Expression);
+                expect(structure.right.subType).to.equal('value');
+                expect(structure.right.value).to.equal(1);
+            }
+        }
     });
 });